Adds new meta extension, adds commandInfo & version to it
