+++ src/domain/runtime-extensions.test.js
@@ -6,5 +6,5 @@ test('loads the core extensions in the right order', t => {
   const r = new Runtime()
   const list = pipe(pluck('name'), join(', '))(r.extensions)
 
-  t.is(list, 'core, strings, print, template, filesystem, semver, system, http, prompt, patching')
+  t.is(list, 'meta, strings, print, template, filesystem, semver, system, http, prompt, patching')
 })
