+++ src/utils/print-help.js
@@ -1,54 +1,5 @@
 const print = require('./print')
-const { pipe, map, sortBy, prop, propEq, reject, replace, unnest, equals } = require('ramda')
-
-/**
- * Is this a hidden command?
- */
-const isHidden = propEq('hidden', true)
-
-/**
- * Gets the list of plugins.
- *
- * @param {RunContext} context     The context
- * @param {Plugin[]} plugins       The plugins holding the commands
- * @param {string[]} commandRoot   Optional, only show commands with this root
- * @return {[string, string]}
- */
-function getListOfPluginCommands (context, plugins, commandRoot) {
-  return pipe(
-    reject(isHidden),
-    sortBy(prop('name')),
-    map(p => getListOfCommands(context, p, commandRoot)),
-    unnest
-  )(plugins)
-}
-
-/**
- * Gets the list of commands for the given plugin.
- *
- * @param {RunContext} context     The context
- * @param {Plugin} plugin          The plugins holding the commands
- * @param {string[]} commandRoot   Optional, only show commands with this root
- * @return {[string, string]}
- */
-function getListOfCommands (context, plugin, commandRoot) {
-  return pipe(
-    reject(isHidden),
-    reject(command => {
-      if (!commandRoot) {
-        return false
-      }
-      return !equals(command.commandPath.slice(0, commandRoot.length), commandRoot)
-    }),
-    map(command => {
-      const alias = command.hasAlias() ? `(${command.aliases.join(', ')})` : ''
-      return [
-        `${command.commandPath.join(' ')} ${alias}`,
-        replace('$BRAND', context.runtime.brand, command.description || '-')
-      ]
-    })
-  )(plugin.commands)
-}
+const { commandInfo } = require('./command-info')
 
 /**
  * Prints the list of commands.
@@ -66,7 +17,7 @@ function printCommands (context, commandRoot) {
     printPlugins = [context.plugin]
   }
 
-  const data = getListOfPluginCommands(context, printPlugins, commandRoot)
+  const data = commandInfo(context, printPlugins, commandRoot)
 
   print.newline() // a spacer
   print.table(data) // the data
@@ -74,7 +25,7 @@ function printCommands (context, commandRoot) {
 
 function printHelp (context) {
   const { print, runtime: { brand } } = context
-  print.info(`${brand} version ${context.version()}`)
+  print.info(`${brand} version ${context.meta.version()}`)
   print.printCommands(context)
 }
